functions: split the code
